feat: show SSO status in user details
Closes #40
