+++ src/Resources/app/administration/src/app/components/heptacom-admin-open-auth-url-clients/snippet/de-DE.json
@@ -1,0 +1,19 @@
+{
+    "heptacom-admin-open-auth-url-clients": {
+        "userKeys": {
+            "actions": {
+                "connect": "Verbinden",
+                "revoke": "Verbindung trennen"
+            },
+            "columns": {
+                "createdAt": "Verbunden am",
+                "provider": "Anbieter"
+            },
+            "states": {
+                "connected": "Verbunden",
+                "notConnected": "Nicht verbunden"
+            }
+        },
+        "title": "Single Sign-on"
+    }
+}
