+++ src/Resources/app/administration/src/app/index.js
@@ -2,3 +2,4 @@ const { Component } = Shopware;
 
 Component.extend('heptacom-admin-open-auth-url-field', 'sw-url-field', () => import('./components/heptacom-admin-open-auth-url-field'));
 Component.register('heptacom-admin-open-auth-user-confirm-login', () => import('./components/heptacom-admin-open-auth-user-confirm-login'));
+Component.register('heptacom-admin-open-auth-url-clients', () => import('./components/heptacom-admin-open-auth-url-clients'));
