--- src/Resources/app/administration/src/extension/snippet/de-DE.json
@@ -1,17 +1,0 @@
-{
-    "sw-profile-index": {
-        "heptacomAdminOpenAuth": {
-            "userKeys": {
-                "actions": {
-                    "connect": "Verbinden",
-                    "revoke": "Verbindung trennen"
-                },
-                "columns": {
-                    "createdAt": "Verbunden am",
-                    "provider": "Anbieter"
-                }
-            }
-        },
-        "titleHeptacomAdminOpenAuthCard": "Single Sign-on"
-    }
-}
