--- src/Resources/app/administration/src/extension/snippet/en-GB.json
@@ -1,17 +1,0 @@
-{
-    "sw-profile-index": {
-        "heptacomAdminOpenAuth": {
-            "userKeys": {
-                "actions": {
-                    "connect": "Connect",
-                    "revoke": "Disconnect"
-                },
-                "columns": {
-                    "createdAt": "Connected at",
-                    "provider": "Provider"
-                }
-            }
-        },
-        "titleHeptacomAdminOpenAuthCard": "Single Sign-on"
-    }
-}
