+++ src/Resources/app/administration/src/extension/sw-users-permissions-user-detail/index.js
@@ -1,0 +1,7 @@
+import template from './sw-users-permissions-user-detail.html.twig';
+
+const { Component } = Shopware;
+
+Component.override('sw-users-permissions-user-detail', {
+    template,
+});
