+++ src/Resources/app/administration/src/main.js
@@ -1,5 +1,6 @@
 import './app';
 import './extension/sw-profile-index-general';
+import './extension/sw-users-permissions-user-detail';
 import './extension/sw-verify-user-modal';
 import './module/heptacom-admin-open-auth-client';
 import './provider';
